+++ dtool_lookup_server/permission_routes.py
@@ -1,6 +1,5 @@
 from flask import (
     abort,
-    request,
     jsonify,
 )
 from flask_jwt_extended import (
@@ -9,8 +8,8 @@ from flask_jwt_extended import (
 )
 from flask_smorest import Blueprint
 
-from dtool_lookup_server import AuthenticationError
 import dtool_lookup_server.utils
+from dtool_lookup_server import AuthenticationError
 from dtool_lookup_server.schemas import BaseUriSchema, UriPermissionSchema
 
 bp = Blueprint("permissions", __name__, url_prefix="/admin/permission")
@@ -19,7 +18,7 @@ bp = Blueprint("permissions", __name__, url_prefix="/admin/permission")
 @bp.route("/info", methods=["POST"])
 @bp.arguments(BaseUriSchema)
 @jwt_required()
-def permission_info(base_uri: BaseUriSchema):
+def permission_info(data: BaseUriSchema):
     """Get information about the permissions on a base URI.
 
     The user needs to be admin.
@@ -36,7 +35,6 @@ def permission_info(base_uri: BaseUriSchema):
     if not user.is_admin:
         abort(404)
 
-    data = request.get_json()
     base_uri = data["base_uri"]
 
     return jsonify(dtool_lookup_server.utils.get_permission_info(base_uri))
@@ -65,7 +63,6 @@ def update_on_base_uri(permissions: UriPermissionSchema):
 
     # TODO: is it safe to pass this information straight through without
     #       validation?
-    permissions = request.get_json()
     dtool_lookup_server.utils.update_permissions(permissions)
 
     return "", 201